backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
cryptography
supabase==2.3.5
python-jose
